+++ anyblok_marshmallow/__init__.py
@@ -7,6 +7,6 @@
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
 
-from .schema import ModelSchema  # noqa
+from .schema import ModelSchema, PostLoadSchema  # noqa
 from .exceptions import RegistryNotFound  # noqa
 from .fields import Nested, File  # noqa